+++ handlers/iiifhandler/iiifparser/IIIFRegion.h
@@ -17,8 +17,8 @@ namespace cserve {
     private:
         CoordType coord_type;
         float rx, ry, rw, rh;
-        int x, y;
-        size_t w, h;
+        int32_t x, y;
+        uint32_t w, h;
         bool canonical_ok;
 
     public:
@@ -28,7 +28,7 @@ namespace cserve {
             canonical_ok = false;
         }
 
-        inline IIIFRegion(int x, int y, size_t w, size_t h) : x(0), y(0), w(0), h(0) {
+        inline IIIFRegion(int32_t x, int32_t y, uint32_t w, uint32_t h) : x(0), y(0), w(0), h(0) {
             coord_type = COORDS;
             rx = (float) x;
             ry = (float) y;
@@ -55,7 +55,7 @@ namespace cserve {
          *
          * \returns CoordType
          */
-        CoordType crop_coords(size_t nx, size_t ny, int &p_x, int &p_y, size_t &p_w, size_t &p_h);
+        CoordType crop_coords(uint32_t nx, uint32_t ny, int32_t &p_x, int32_t &p_y, uint32_t &p_w, uint32_t &p_h);
 
         /*!
          * Returns the canoncial IIIF string for the given region
